pyproject.toml
[project]
name = "terrabridge-mcp"
version = "0.1.0"
description = "on-premise agentic ai with mcp for dms/cassandra and iso 27001 compliance monitoring"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.2",
    "httpx[http2]>=0.28.1",
    "pydantic>=2.12.5",
    "cassandra-driver>=3.29.3",
    "python-dotenv>=1.2.1",
    "mlx>=0.30.0; sys_platform == 'darwin'",
    "mlx-lm>=0.28.3; sys_platform == 'darwin'",
]

# note: tensorrt is an alternative to vllm on linux (mutually exclusive)
[project.optional-dependencies]
tensorrt = [
    "tensorrt-llm>=1.2.0rc4",
    "etcd3 @ git+https://github.com/kragniz/python-etcd3.git@e58a899579ba416449c4e225b61f039457c8072a",
]
vllm = ["vllm>=0.12.0; sys_platform == 'linux'"]

[tool.uv]
conflicts = [
    [
      { extra = "tensorrt" },
      { extra = "vllm" },
    ],
]

[project.scripts]
mcp-server = "protocol_layer.server:main"
agent = "orchestration_layer.agent:main"
server = "serving_layer.server:main"
# Compliance monitoring
compliance-sensor = "sensor_layer.broadcaster:main"
compliance-scan = "sensor_layer.scanner:main"

[build-system]
requires = ["hatchling", "editables>=0.5"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["protocol_layer", "orchestration_layer", "serving_layer", "sensor_layer"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.extra-build-dependencies]
"cassandra_driver-3.29.3" = ["setuptools"]

[dependency-groups]
test = ["pytest>=9.0.1"]
dev = ["ruff>=0.14.7", "pytest>=9.0.1"]
